Added pages as slides, worked on hero editing
Now each main page, "Crafting (not yet implemented), Heroes, Cards, Quests" is inside a slider that can be easily navigated.

Finished some functionality on the hero preview, now user can drag and drop cards to change equipment and actions

On heroes instead of null equipment, if no equipment for a slot is found, a basic equipment with no stats is equipped.